Added command history logging to all cli commands
