+++ yo_jenkins/cli/cli_server.py
@@ -7,10 +7,11 @@ import sys
 
 import click
 from yo_jenkins.Docker import DockerJenkinsServer
-from yo_jenkins.Utility.utility import (get_project_dir, get_resource_path, log_to_history)
+from yo_jenkins.Utility.utility import get_project_dir, get_resource_path
 from yo_jenkins.YoJenkins import Auth, YoJenkins
 
-from . import cli_utility as cu
+from yo_jenkins.cli import cli_utility as cu
+from yo_jenkins.cli.cli_utility import log_to_history
 
 # Getting the logger reference
 logger = logging.getLogger()
